requirements.txt
flask
azure-cosmos